+++ cffi_dist/go.mod
@@ -18,3 +18,5 @@ require (
 	golang.org/x/sys v0.4.0 // indirect
 	golang.org/x/text v0.6.0 // indirect
 )
+
+// replace github.com/bogdanfinn/tls-client => ../
