+++ cffi_src/factory.go
@@ -5,6 +5,7 @@ import (
 	"bytes"
 	"encoding/base64"
 	"fmt"
+	"github.com/bogdanfinn/tls-client/profiles"
 	"io"
 	"net"
 	"os"
@@ -256,7 +257,7 @@ func getTlsClient(requestInput RequestInput, sessionId string, withSession bool)
 		return modifiedClient, nil
 	}
 
-	clientProfile := tls_client.DefaultClientProfile
+	clientProfile := profiles.DefaultClientProfile
 
 	if requestInput.CustomTlsClient != nil {
 		clientHelloId, h2Settings, h2SettingsOrder, pseudoHeaderOrder, connectionFlow, priorityFrames, headerPriority, err := getCustomTlsClientProfile(requestInput.CustomTlsClient)
@@ -264,7 +265,7 @@ func getTlsClient(requestInput RequestInput, sessionId string, withSession bool)
 			return nil, fmt.Errorf("can not build http client out of custom tls client information: %w", err)
 		}
 
-		clientProfile = tls_client.NewClientProfile(clientHelloId, h2Settings, h2SettingsOrder, pseudoHeaderOrder, connectionFlow, priorityFrames, headerPriority)
+		clientProfile = profiles.NewClientProfile(clientHelloId, h2Settings, h2SettingsOrder, pseudoHeaderOrder, connectionFlow, priorityFrames, headerPriority)
 	}
 
 	if tlsClientIdentifier != "" {
@@ -444,11 +445,11 @@ func getCustomTlsClientProfile(customClientDefinition *CustomTlsClient) (tls.Cli
 	return clientHelloId, resolvedH2Settings, resolvedH2SettingsOrder, pseudoHeaderOrder, connectionFlow, priorityFrames, headerPriority, nil
 }
 
-func getTlsClientProfile(tlsClientIdentifier string) tls_client.ClientProfile {
-	tlsClientProfile, ok := tls_client.MappedTLSClients[tlsClientIdentifier]
+func getTlsClientProfile(tlsClientIdentifier string) profiles.ClientProfile {
+	tlsClientProfile, ok := profiles.MappedTLSClients[tlsClientIdentifier]
 
 	if !ok {
-		return tls_client.DefaultClientProfile
+		return profiles.DefaultClientProfile
 	}
 
 	return tlsClientProfile
