+++ client_options.go
@@ -3,6 +3,7 @@ package tls_client
 import (
 	"crypto/x509"
 	"fmt"
+	"github.com/bogdanfinn/tls-client/profiles"
 	"net"
 	"time"
 
@@ -43,7 +44,7 @@ type httpClientConfig struct {
 	serverNameOverwrite         string
 	transportOptions            *TransportOptions
 	cookieJar                   http.CookieJar
-	clientProfile               ClientProfile
+	clientProfile               profiles.ClientProfile
 	withRandomTlsExtensionOrder bool
 	forceHttp1                  bool
 	timeout                     time.Duration
@@ -198,7 +199,7 @@ func WithForceHttp1() HttpClientOption {
 }
 
 // WithClientProfile configures a TLS client to use the specified client profile.
-func WithClientProfile(clientProfile ClientProfile) HttpClientOption {
+func WithClientProfile(clientProfile profiles.ClientProfile) HttpClientOption {
 	return func(config *httpClientConfig) {
 		config.clientProfile = clientProfile
 	}
