+++ example/main.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
+	"github.com/bogdanfinn/tls-client/profiles"
 	"io"
 	"log"
 	"net/url"
@@ -44,7 +45,7 @@ func sslPinning() {
 
 	options := []tls_client.HttpClientOption{
 		tls_client.WithTimeoutSeconds(60),
-		tls_client.WithClientProfile(tls_client.Chrome_108),
+		tls_client.WithClientProfile(profiles.Chrome_108),
 		tls_client.WithRandomTLSExtensionOrder(),
 		tls_client.WithCookieJar(jar),
 		tls_client.WithCertificatePinning(pins, tls_client.DefaultBadPinHandler),
@@ -106,7 +107,7 @@ func requestToppsAsChrome107Client() {
 
 	options := []tls_client.HttpClientOption{
 		tls_client.WithTimeoutSeconds(30),
-		tls_client.WithClientProfile(tls_client.Chrome_107),
+		tls_client.WithClientProfile(profiles.Chrome_107),
 		tls_client.WithDebug(),
 		// tls_client.WithProxyUrl("http://user:pass@host:port"),
 		// tls_client.WithNotFollowRedirects(),
@@ -181,7 +182,7 @@ func requestToppsAsChrome107Client() {
 func postAsTlsClient() {
 	options := []tls_client.HttpClientOption{
 		tls_client.WithTimeoutSeconds(30),
-		tls_client.WithClientProfile(tls_client.Chrome_107),
+		tls_client.WithClientProfile(profiles.Chrome_107),
 	}
 
 	client, err := tls_client.NewHttpClient(tls_client.NewNoopLogger(), options...)
@@ -229,7 +230,7 @@ func postAsTlsClient() {
 func requestWithFollowRedirectSwitch() {
 	options := []tls_client.HttpClientOption{
 		tls_client.WithTimeoutSeconds(30),
-		tls_client.WithClientProfile(tls_client.Chrome_107),
+		tls_client.WithClientProfile(profiles.Chrome_107),
 		tls_client.WithNotFollowRedirects(),
 	}
 
@@ -305,7 +306,7 @@ func requestWithFollowRedirectSwitch() {
 func downloadImageWithTlsClient() {
 	options := []tls_client.HttpClientOption{
 		tls_client.WithTimeoutSeconds(30),
-		tls_client.WithClientProfile(tls_client.Chrome_107),
+		tls_client.WithClientProfile(profiles.Chrome_107),
 		tls_client.WithNotFollowRedirects(),
 	}
 
@@ -362,7 +363,7 @@ func downloadImageWithTlsClient() {
 func rotateProxiesOnClient() {
 	options := []tls_client.HttpClientOption{
 		tls_client.WithTimeoutSeconds(30),
-		tls_client.WithClientProfile(tls_client.Chrome_107),
+		tls_client.WithClientProfile(profiles.Chrome_107),
 		tls_client.WithProxyUrl("http://user:pass@host:port"),
 	}
 
@@ -561,7 +562,7 @@ func requestWithCustomClient() {
 		}, nil
 	}
 
-	customClientProfile := tls_client.NewClientProfile(tls.ClientHelloID{
+	customClientProfile := profiles.NewClientProfile(tls.ClientHelloID{
 		Client:      "MyCustomProfile",
 		Version:     "1",
 		Seed:        nil,
