+++ profiles/contributed_browser_profiles.go
@@ -1,0 +1,1 @@
+package profiles
