+++ profiles/contributed_custom_profiles.go
@@ -1,10 +1,9 @@
-package tls_client
+package profiles
 
 import (
-	"math"
-
 	"github.com/bogdanfinn/fhttp/http2"
 	tls "github.com/bogdanfinn/utls"
+	"math"
 )
 
 var ZalandoAndroidMobile = ClientProfile{
@@ -99,208 +98,6 @@ var ZalandoAndroidMobile = ClientProfile{
 	connectionFlow: 15663105,
 }
 
-var ZalandoIosMobile = ClientProfile{
-	clientHelloId: tls.ClientHelloID{
-		Client:  "ZalandoIosCustom",
-		Version: "1",
-		Seed:    nil,
-		SpecFactory: func() (tls.ClientHelloSpec, error) {
-			return tls.ClientHelloSpec{
-				CipherSuites: []uint16{
-					tls.GREASE_PLACEHOLDER,
-					tls.TLS_AES_128_GCM_SHA256,
-					tls.TLS_AES_256_GCM_SHA384,
-					tls.TLS_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-					tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-					tls.TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA,
-					tls.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA,
-					tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA,
-				},
-				CompressionMethods: []byte{
-					tls.CompressionNone,
-				},
-				Extensions: []tls.TLSExtension{
-					&tls.UtlsGREASEExtension{},
-					&tls.SNIExtension{},
-					&tls.UtlsExtendedMasterSecretExtension{},
-					&tls.RenegotiationInfoExtension{Renegotiation: tls.RenegotiateOnceAsClient},
-					&tls.SupportedCurvesExtension{Curves: []tls.CurveID{
-						tls.CurveID(tls.GREASE_PLACEHOLDER),
-						tls.X25519,
-						tls.CurveP256,
-						tls.CurveP384,
-						tls.CurveP521,
-					}},
-					&tls.SupportedPointsExtension{SupportedPoints: []byte{
-						tls.PointFormatUncompressed,
-					}},
-					&tls.ALPNExtension{AlpnProtocols: []string{"h2", "http/1.1"}},
-					&tls.StatusRequestExtension{},
-					&tls.SignatureAlgorithmsExtension{SupportedSignatureAlgorithms: []tls.SignatureScheme{
-						tls.ECDSAWithP256AndSHA256,
-						tls.PSSWithSHA256,
-						tls.PKCS1WithSHA256,
-						tls.ECDSAWithP384AndSHA384,
-						tls.ECDSAWithSHA1,
-						tls.PSSWithSHA384,
-						tls.PSSWithSHA384,
-						tls.PKCS1WithSHA384,
-						tls.PSSWithSHA512,
-						tls.PKCS1WithSHA512,
-						tls.PKCS1WithSHA1,
-					}},
-					&tls.SCTExtension{},
-					&tls.KeyShareExtension{KeyShares: []tls.KeyShare{
-						{Group: tls.CurveID(tls.GREASE_PLACEHOLDER), Data: []byte{0}},
-						{Group: tls.X25519},
-					}},
-					&tls.PSKKeyExchangeModesExtension{Modes: []uint8{
-						tls.PskModeDHE,
-					}},
-					&tls.SupportedVersionsExtension{Versions: []uint16{
-						tls.GREASE_PLACEHOLDER,
-						tls.VersionTLS13,
-						tls.VersionTLS12,
-					}},
-					&tls.UtlsCompressCertExtension{Algorithms: []tls.CertCompressionAlgo{
-						tls.CertCompressionZlib,
-					}},
-					&tls.UtlsGREASEExtension{},
-					&tls.UtlsPaddingExtension{GetPaddingLen: tls.BoringPaddingStyle},
-				},
-			}, nil
-		},
-	},
-	settings: map[http2.SettingID]uint32{
-		http2.SettingHeaderTableSize:      4096,
-		http2.SettingMaxConcurrentStreams: 100,
-		http2.SettingInitialWindowSize:    2097152,
-		http2.SettingMaxFrameSize:         16384,
-		http2.SettingMaxHeaderListSize:    math.MaxUint32,
-	},
-	settingsOrder: []http2.SettingID{
-		http2.SettingHeaderTableSize,
-		http2.SettingMaxConcurrentStreams,
-		http2.SettingInitialWindowSize,
-		http2.SettingMaxFrameSize,
-		http2.SettingMaxHeaderListSize,
-	},
-	pseudoHeaderOrder: []string{
-		":method",
-		":path",
-		":authority",
-		":scheme",
-	},
-	connectionFlow: 15663105,
-}
-
-var NikeIosMobile = ClientProfile{
-	clientHelloId: tls.ClientHelloID{
-		Client:  "NikeIosCustom",
-		Version: "1",
-		Seed:    nil,
-		SpecFactory: func() (tls.ClientHelloSpec, error) {
-			return tls.ClientHelloSpec{
-				CipherSuites: []uint16{
-					tls.GREASE_PLACEHOLDER,
-					tls.TLS_AES_128_GCM_SHA256,
-					tls.TLS_AES_256_GCM_SHA384,
-					tls.TLS_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-					tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-					tls.TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA,
-					tls.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA,
-					tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA,
-				},
-				CompressionMethods: []uint8{
-					tls.CompressionNone,
-				},
-				Extensions: []tls.TLSExtension{
-					&tls.UtlsGREASEExtension{},
-					&tls.SNIExtension{},
-					&tls.UtlsExtendedMasterSecretExtension{},
-					&tls.RenegotiationInfoExtension{Renegotiation: tls.RenegotiateOnceAsClient},
-					&tls.SupportedCurvesExtension{Curves: []tls.CurveID{
-						tls.CurveID(tls.GREASE_PLACEHOLDER),
-						tls.X25519,
-						tls.CurveP256,
-						tls.CurveP384,
-						tls.CurveP521,
-					}},
-					&tls.SupportedPointsExtension{SupportedPoints: []byte{
-						tls.PointFormatUncompressed,
-					}},
-					&tls.ALPNExtension{AlpnProtocols: []string{"h2", "http/1.1"}},
-					&tls.StatusRequestExtension{},
-					&tls.SignatureAlgorithmsExtension{SupportedSignatureAlgorithms: []tls.SignatureScheme{
-						tls.ECDSAWithP256AndSHA256,
-						tls.PSSWithSHA256,
-						tls.PKCS1WithSHA256,
-						tls.ECDSAWithP384AndSHA384,
-						tls.ECDSAWithSHA1,
-						tls.PSSWithSHA384,
-						tls.PSSWithSHA384,
-						tls.PKCS1WithSHA384,
-						tls.PSSWithSHA512,
-						tls.PKCS1WithSHA512,
-						tls.PKCS1WithSHA1,
-					}},
-					&tls.SCTExtension{},
-					&tls.KeyShareExtension{KeyShares: []tls.KeyShare{
-						{Group: tls.CurveID(tls.GREASE_PLACEHOLDER), Data: []byte{0}},
-						{Group: tls.X25519},
-					}},
-					&tls.PSKKeyExchangeModesExtension{Modes: []uint8{
-						tls.PskModeDHE,
-					}},
-					&tls.SupportedVersionsExtension{Versions: []uint16{
-						tls.GREASE_PLACEHOLDER,
-						tls.VersionTLS13,
-						tls.VersionTLS12,
-					}},
-					&tls.UtlsCompressCertExtension{Algorithms: []tls.CertCompressionAlgo{
-						tls.CertCompressionZlib,
-					}},
-					&tls.UtlsGREASEExtension{},
-					&tls.UtlsPaddingExtension{GetPaddingLen: tls.BoringPaddingStyle},
-				},
-			}, nil
-		},
-	},
-	settings: map[http2.SettingID]uint32{
-		http2.SettingHeaderTableSize:      4096,
-		http2.SettingMaxConcurrentStreams: 100,
-		http2.SettingInitialWindowSize:    2097152,
-		http2.SettingMaxFrameSize:         16384,
-		http2.SettingMaxHeaderListSize:    math.MaxUint32,
-	},
-	settingsOrder: []http2.SettingID{
-		http2.SettingHeaderTableSize,
-		http2.SettingMaxConcurrentStreams,
-		http2.SettingInitialWindowSize,
-		http2.SettingMaxFrameSize,
-		http2.SettingMaxHeaderListSize,
-	},
-	pseudoHeaderOrder: []string{
-		":method",
-		":scheme",
-		":path",
-		":authority",
-	},
-	connectionFlow: 15663105,
-}
-
 var NikeAndroidMobile = ClientProfile{
 	clientHelloId: tls.ClientHelloID{
 		Client:  "NikeAndroidCustom",
@@ -389,288 +186,6 @@ var NikeAndroidMobile = ClientProfile{
 		":path",
 		":authority",
 		":scheme",
-	},
-	connectionFlow: 15663105,
-}
-
-var CloudflareCustom = ClientProfile{
-	clientHelloId: tls.ClientHelloID{
-		Client:  "CloudflareCustom",
-		Version: "1",
-		Seed:    nil,
-		SpecFactory: func() (tls.ClientHelloSpec, error) {
-			return tls.ClientHelloSpec{
-				CipherSuites: []uint16{
-					tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA,
-					tls.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA,
-					tls.FAKE_TLS_EMPTY_RENEGOTIATION_INFO_SCSV,
-				},
-				CompressionMethods: []uint8{
-					tls.CompressionNone,
-				},
-				Extensions: []tls.TLSExtension{
-					&tls.SNIExtension{},
-					&tls.SupportedPointsExtension{SupportedPoints: []uint8{
-						tls.PointFormatUncompressed,
-						1, // ansiX962_compressed_prime
-						2, // ansiX962_compressed_char2
-					}},
-					&tls.SupportedCurvesExtension{Curves: []tls.CurveID{
-						tls.CurveID(0x0017),
-					}},
-					&tls.SessionTicketExtension{},
-					// due to that we do not care about http2 frame settings
-					&tls.ALPNExtension{AlpnProtocols: []string{"http/1.1"}},
-					&tls.GenericExtension{Id: 22}, // encrypt_then_mac
-					&tls.UtlsExtendedMasterSecretExtension{},
-					&tls.SignatureAlgorithmsExtension{SupportedSignatureAlgorithms: []tls.SignatureScheme{
-						tls.ECDSAWithP256AndSHA256,
-						tls.ECDSAWithP384AndSHA384,
-						tls.ECDSAWithP521AndSHA512,
-						0x0807,
-						0x0808,
-						0x0809,
-						0x080a,
-						0x080b,
-						tls.PSSWithSHA256,
-						tls.PSSWithSHA384,
-						tls.PSSWithSHA512,
-						tls.PKCS1WithSHA256,
-						tls.PKCS1WithSHA384,
-						tls.PKCS1WithSHA512,
-						0x0303,
-						0x0203,
-						0x0301,
-						0x0201,
-						0x0302,
-						0x0202,
-						0x0402,
-						0x0502,
-						0x0602,
-					}},
-				},
-			}, nil
-		},
-	},
-
-	//actually the h2 settings are not relevant, because this client does only support http1
-	settings: map[http2.SettingID]uint32{
-		http2.SettingHeaderTableSize:      4096,
-		http2.SettingMaxConcurrentStreams: math.MaxUint32,
-		http2.SettingInitialWindowSize:    16777216,
-		http2.SettingMaxFrameSize:         16384,
-		http2.SettingMaxHeaderListSize:    math.MaxUint32,
-	},
-	settingsOrder: []http2.SettingID{
-		http2.SettingHeaderTableSize,
-		http2.SettingMaxConcurrentStreams,
-		http2.SettingInitialWindowSize,
-		http2.SettingMaxFrameSize,
-		http2.SettingMaxHeaderListSize,
-	},
-	pseudoHeaderOrder: []string{
-		":method",
-		":path",
-		":authority",
-		":scheme",
-	},
-	connectionFlow: 15663105,
-}
-
-var MMSIos = ClientProfile{
-	clientHelloId: tls.ClientHelloID{
-		Client:  "MMSIos",
-		Version: "1",
-		Seed:    nil,
-		SpecFactory: func() (tls.ClientHelloSpec, error) {
-			return tls.ClientHelloSpec{
-				CipherSuites: []uint16{
-					0x1303,
-					0x1301,
-					0x1302,
-					0xcca9,
-					0xcca8,
-					0xc02b,
-					0xc02f,
-					0xc02c,
-					0xc030,
-					0xc009,
-					0xc013,
-					0xc00a,
-					0xc014,
-					0x009c,
-					0x009d,
-					0x002f,
-					0x0035,
-					0x000a,
-				},
-				CompressionMethods: []uint8{
-					tls.CompressionNone,
-				},
-				Extensions: []tls.TLSExtension{
-					&tls.SNIExtension{},
-					&tls.UtlsExtendedMasterSecretExtension{},
-					&tls.RenegotiationInfoExtension{Renegotiation: tls.RenegotiateOnceAsClient},
-					&tls.SupportedCurvesExtension{Curves: []tls.CurveID{
-						tls.CurveID(0x001d),
-						tls.CurveID(0x0017),
-						tls.CurveID(0x0018),
-					}},
-					&tls.SupportedPointsExtension{SupportedPoints: []uint8{
-						tls.PointFormatUncompressed,
-					}},
-					&tls.SessionTicketExtension{},
-					&tls.SignatureAlgorithmsExtension{SupportedSignatureAlgorithms: []tls.SignatureScheme{
-						0x0403,
-						0x0804,
-						0x0401,
-						0x0503,
-						0x0805,
-						0x0501,
-						0x0806,
-						0x0601,
-						0x0201,
-					}},
-					&tls.KeyShareExtension{KeyShares: []tls.KeyShare{
-						{Group: tls.X25519},
-					}},
-					&tls.PSKKeyExchangeModesExtension{Modes: []uint8{
-						tls.PskModeDHE,
-					}},
-					&tls.SupportedVersionsExtension{Versions: []uint16{
-						tls.VersionTLS13,
-						tls.VersionTLS12,
-					}},
-				},
-			}, nil
-		},
-	},
-	settings: map[http2.SettingID]uint32{
-		http2.SettingHeaderTableSize:      4096,
-		http2.SettingEnablePush:           1,
-		http2.SettingMaxConcurrentStreams: 100,
-		http2.SettingInitialWindowSize:    2097152,
-		http2.SettingMaxFrameSize:         16384,
-		http2.SettingMaxHeaderListSize:    math.MaxUint32,
-	},
-	settingsOrder: []http2.SettingID{
-		http2.SettingHeaderTableSize,
-		http2.SettingEnablePush,
-		http2.SettingMaxConcurrentStreams,
-		http2.SettingInitialWindowSize,
-		http2.SettingMaxFrameSize,
-		http2.SettingMaxHeaderListSize,
-	},
-	pseudoHeaderOrder: []string{
-		":method",
-		":scheme",
-		":path",
-		":authority",
-	},
-	connectionFlow: 15663105,
-}
-
-var MeshIos = ClientProfile{
-	clientHelloId: tls.ClientHelloID{
-		Client:  "MeshIos",
-		Version: "1",
-		Seed:    nil,
-		SpecFactory: func() (tls.ClientHelloSpec, error) {
-			return tls.ClientHelloSpec{
-				CipherSuites: []uint16{
-					tls.GREASE_PLACEHOLDER,
-					0x1301,
-					0x1302,
-					0x1303,
-					0xc02c,
-					0xc02b,
-					0xcca9,
-					0xc030,
-					0xc02f,
-					0xcca8,
-					0xc00a,
-					0xc009,
-					0xc014,
-					0xc013,
-				},
-				CompressionMethods: []uint8{
-					tls.CompressionNone,
-				},
-				Extensions: []tls.TLSExtension{
-					&tls.UtlsGREASEExtension{},
-					&tls.SNIExtension{},
-					&tls.UtlsExtendedMasterSecretExtension{},
-					&tls.RenegotiationInfoExtension{Renegotiation: tls.RenegotiateOnceAsClient},
-					&tls.SupportedCurvesExtension{Curves: []tls.CurveID{
-						tls.CurveID(tls.GREASE_PLACEHOLDER),
-						tls.CurveID(0x001d),
-						tls.CurveID(0x0017),
-						tls.CurveID(0x0018),
-						tls.CurveID(0x0019),
-					}},
-					&tls.SupportedPointsExtension{SupportedPoints: []uint8{
-						tls.PointFormatUncompressed,
-					}},
-					&tls.ALPNExtension{AlpnProtocols: []string{"h2", "http/1.1"}},
-					&tls.StatusRequestExtension{},
-					&tls.SignatureAlgorithmsExtension{SupportedSignatureAlgorithms: []tls.SignatureScheme{
-						0x0403,
-						0x0804,
-						0x0401,
-						0x0503,
-						0x0203,
-						0x0805,
-						0x0805,
-						0x0501,
-						0x0806,
-						0x0601,
-						0x0201,
-					}},
-					&tls.SCTExtension{},
-					&tls.KeyShareExtension{KeyShares: []tls.KeyShare{
-						{Group: tls.CurveID(tls.GREASE_PLACEHOLDER), Data: []byte{0}},
-						{Group: tls.X25519},
-					}},
-					&tls.PSKKeyExchangeModesExtension{Modes: []uint8{
-						tls.PskModeDHE,
-					}},
-					&tls.SupportedVersionsExtension{Versions: []uint16{
-						tls.GREASE_PLACEHOLDER,
-						tls.VersionTLS13,
-						tls.VersionTLS12,
-					}},
-					&tls.UtlsCompressCertExtension{Algorithms: []tls.CertCompressionAlgo{
-						tls.CertCompressionZlib,
-					}},
-					&tls.UtlsGREASEExtension{},
-					&tls.UtlsPaddingExtension{GetPaddingLen: tls.BoringPaddingStyle},
-				},
-			}, nil
-		},
-	},
-	settings: map[http2.SettingID]uint32{
-		http2.SettingHeaderTableSize:      4096,
-		http2.SettingEnablePush:           1,
-		http2.SettingMaxConcurrentStreams: 100,
-		http2.SettingInitialWindowSize:    2097152,
-		http2.SettingMaxFrameSize:         16384,
-		http2.SettingMaxHeaderListSize:    math.MaxUint32,
-	},
-	settingsOrder: []http2.SettingID{
-		http2.SettingHeaderTableSize,
-		http2.SettingEnablePush,
-		http2.SettingMaxConcurrentStreams,
-		http2.SettingInitialWindowSize,
-		http2.SettingMaxFrameSize,
-		http2.SettingMaxHeaderListSize,
-	},
-	pseudoHeaderOrder: []string{
-		":method",
-		":scheme",
-		":path",
-		":authority",
 	},
 	connectionFlow: 15663105,
 }
@@ -949,109 +464,6 @@ var MeshAndroid2 = ClientProfile{
 		":authority",
 		":scheme",
 		":path",
-	},
-	connectionFlow: 15663105,
-}
-
-var ConfirmedIos = ClientProfile{
-	clientHelloId: tls.ClientHelloID{
-		Client:  "ConfirmedIos",
-		Version: "1",
-		Seed:    nil,
-		SpecFactory: func() (tls.ClientHelloSpec, error) {
-			return tls.ClientHelloSpec{
-				CipherSuites: []uint16{
-					tls.GREASE_PLACEHOLDER,
-					tls.TLS_AES_128_GCM_SHA256,
-					tls.TLS_AES_256_GCM_SHA384,
-					tls.TLS_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-					tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
-					tls.TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA,
-					tls.TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA,
-					tls.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA,
-					tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA,
-				},
-				CompressionMethods: []uint8{
-					tls.CompressionNone,
-				},
-				Extensions: []tls.TLSExtension{
-					&tls.UtlsGREASEExtension{},
-					&tls.SNIExtension{},
-					&tls.UtlsExtendedMasterSecretExtension{},
-					&tls.RenegotiationInfoExtension{Renegotiation: tls.RenegotiateOnceAsClient},
-					&tls.SupportedCurvesExtension{Curves: []tls.CurveID{
-						tls.CurveID(tls.GREASE_PLACEHOLDER),
-						tls.X25519,
-						tls.CurveP256,
-						tls.CurveP384,
-						tls.CurveP521,
-					}},
-					&tls.SupportedPointsExtension{SupportedPoints: []uint8{
-						tls.PointFormatUncompressed,
-					}},
-					&tls.ALPNExtension{AlpnProtocols: []string{"h2", "http/1.1"}},
-					&tls.StatusRequestExtension{},
-					&tls.SignatureAlgorithmsExtension{SupportedSignatureAlgorithms: []tls.SignatureScheme{
-						tls.ECDSAWithP256AndSHA256,
-						tls.PSSWithSHA256,
-						tls.PKCS1WithSHA256,
-						tls.ECDSAWithP384AndSHA384,
-						tls.ECDSAWithSHA1,
-						tls.PSSWithSHA384,
-						tls.PSSWithSHA384,
-						tls.PKCS1WithSHA384,
-						tls.PSSWithSHA512,
-						tls.PKCS1WithSHA512,
-						tls.PKCS1WithSHA1,
-					}},
-					&tls.SCTExtension{},
-					&tls.KeyShareExtension{KeyShares: []tls.KeyShare{
-						{Group: tls.CurveID(tls.GREASE_PLACEHOLDER), Data: []byte{0}},
-						{Group: tls.X25519},
-					}},
-					&tls.PSKKeyExchangeModesExtension{Modes: []uint8{
-						tls.PskModeDHE,
-					}},
-					&tls.SupportedVersionsExtension{Versions: []uint16{
-						tls.GREASE_PLACEHOLDER,
-						tls.VersionTLS13,
-						tls.VersionTLS12,
-					}},
-					&tls.UtlsCompressCertExtension{Algorithms: []tls.CertCompressionAlgo{
-						tls.CertCompressionZlib,
-					}},
-					&tls.UtlsGREASEExtension{},
-					&tls.UtlsPaddingExtension{GetPaddingLen: tls.BoringPaddingStyle},
-				},
-			}, nil
-		},
-	},
-	settings: map[http2.SettingID]uint32{
-		http2.SettingHeaderTableSize:      4096,
-		http2.SettingEnablePush:           1,
-		http2.SettingMaxConcurrentStreams: 100,
-		http2.SettingInitialWindowSize:    2097152,
-		http2.SettingMaxFrameSize:         16384,
-		http2.SettingMaxHeaderListSize:    math.MaxUint32,
-	},
-	settingsOrder: []http2.SettingID{
-		http2.SettingHeaderTableSize,
-		http2.SettingEnablePush,
-		http2.SettingMaxConcurrentStreams,
-		http2.SettingInitialWindowSize,
-		http2.SettingMaxFrameSize,
-		http2.SettingMaxHeaderListSize,
-	},
-	pseudoHeaderOrder: []string{
-		":method",
-		":scheme",
-		":path",
-		":authority",
 	},
 	connectionFlow: 15663105,
 }
