+++ profiles/internal_browser_profiles.go
@@ -1,90 +1,9 @@
-package tls_client
+package profiles
 
 import (
 	"github.com/bogdanfinn/fhttp/http2"
 	tls "github.com/bogdanfinn/utls"
 )
-
-var DefaultClientProfile = Chrome_112
-
-var MappedTLSClients = map[string]ClientProfile{
-	"chrome_103":             Chrome_103,
-	"chrome_104":             Chrome_104,
-	"chrome_105":             Chrome_105,
-	"chrome_106":             Chrome_106,
-	"chrome_107":             Chrome_107,
-	"chrome_108":             Chrome_108,
-	"chrome_109":             Chrome_109,
-	"chrome_110":             Chrome_110,
-	"chrome_111":             Chrome_111,
-	"chrome_112":             Chrome_112,
-	"safari_15_6_1":          Safari_15_6_1,
-	"safari_16_0":            Safari_16_0,
-	"safari_ipad_15_6":       Safari_Ipad_15_6,
-	"safari_ios_15_5":        Safari_IOS_15_5,
-	"safari_ios_15_6":        Safari_IOS_15_6,
-	"safari_ios_16_0":        Safari_IOS_16_0,
-	"firefox_102":            Firefox_102,
-	"firefox_104":            Firefox_104,
-	"firefox_105":            Firefox_105,
-	"firefox_106":            Firefox_106,
-	"firefox_108":            Firefox_108,
-	"firefox_110":            Firefox_110,
-	"opera_89":               Opera_89,
-	"opera_90":               Opera_90,
-	"opera_91":               Opera_91,
-	"zalando_android_mobile": ZalandoAndroidMobile,
-	"zalando_ios_mobile":     ZalandoIosMobile,
-	"nike_ios_mobile":        NikeIosMobile,
-	"nike_android_mobile":    NikeAndroidMobile,
-	"cloudscraper":           CloudflareCustom,
-	"mms_ios":                MMSIos,
-	"mesh_ios":               MeshIos,
-	"mesh_ios_1":             MeshIos,
-	"mesh_ios_2":             MeshIos2,
-	"mesh_android":           MeshAndroid,
-	"mesh_android_1":         MeshAndroid,
-	"mesh_android_2":         MeshAndroid2,
-	"confirmed_ios":          ConfirmedIos,
-	"confirmed_android":      ConfirmedAndroid,
-	"okhttp4_android_7":      Okhttp4Android7,
-	"okhttp4_android_8":      Okhttp4Android8,
-	"okhttp4_android_9":      Okhttp4Android9,
-	"okhttp4_android_10":     Okhttp4Android10,
-	"okhttp4_android_11":     Okhttp4Android11,
-	"okhttp4_android_12":     Okhttp4Android12,
-	"okhttp4_android_13":     Okhttp4Android13,
-}
-
-type ClientProfile struct {
-	clientHelloId     tls.ClientHelloID
-	connectionFlow    uint32
-	headerPriority    *http2.PriorityParam
-	priorities        []http2.Priority
-	pseudoHeaderOrder []string
-	settings          map[http2.SettingID]uint32
-	settingsOrder     []http2.SettingID
-}
-
-func NewClientProfile(clientHelloId tls.ClientHelloID, settings map[http2.SettingID]uint32, settingsOrder []http2.SettingID, pseudoHeaderOrder []string, connectionFlow uint32, priorities []http2.Priority, headerPriority *http2.PriorityParam) ClientProfile {
-	return ClientProfile{
-		clientHelloId:     clientHelloId,
-		settings:          settings,
-		settingsOrder:     settingsOrder,
-		pseudoHeaderOrder: pseudoHeaderOrder,
-		connectionFlow:    connectionFlow,
-		priorities:        priorities,
-		headerPriority:    headerPriority,
-	}
-}
-
-func (c ClientProfile) GetClientHelloSpec() (tls.ClientHelloSpec, error) {
-	return c.clientHelloId.ToSpec()
-}
-
-func (c ClientProfile) GetClientHelloStr() string {
-	return c.clientHelloId.Str()
-}
 
 var Chrome_112 = ClientProfile{
 	clientHelloId: tls.HelloChrome_112,
@@ -442,6 +361,148 @@ var Safari_IOS_15_6 = ClientProfile{
 		":authority",
 	},
 	connectionFlow: 10485760,
+}
+
+var Firefox_117 = ClientProfile{
+	clientHelloId: tls.ClientHelloID{
+		Client:               "Firefox",
+		RandomExtensionOrder: false,
+		Version:              "117",
+		Seed:                 nil,
+		SpecFactory: func() (tls.ClientHelloSpec, error) {
+			return tls.ClientHelloSpec{
+				CipherSuites: []uint16{
+					tls.TLS_AES_128_GCM_SHA256,
+					tls.TLS_CHACHA20_POLY1305_SHA256,
+					tls.TLS_AES_256_GCM_SHA384,
+					tls.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
+					tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
+					tls.TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256,
+					tls.TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256,
+					tls.TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
+					tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
+					tls.TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA,
+					tls.TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA,
+					tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA,
+					tls.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA,
+					tls.TLS_RSA_WITH_AES_128_GCM_SHA256,
+					tls.TLS_RSA_WITH_AES_256_GCM_SHA384,
+					tls.TLS_RSA_WITH_AES_128_CBC_SHA,
+					tls.TLS_RSA_WITH_AES_256_CBC_SHA,
+				},
+				CompressionMethods: []byte{
+					tls.CompressionNone,
+				},
+				Extensions: []tls.TLSExtension{
+					&tls.SNIExtension{},
+					&tls.UtlsExtendedMasterSecretExtension{},
+					&tls.RenegotiationInfoExtension{Renegotiation: tls.RenegotiateOnceAsClient},
+					&tls.SupportedCurvesExtension{[]tls.CurveID{
+						tls.X25519,
+						tls.CurveP256,
+						tls.CurveP384,
+						tls.CurveP521,
+						tls.FAKEFFDHE2048,
+						tls.FAKEFFDHE3072,
+					}},
+					&tls.SupportedPointsExtension{SupportedPoints: []byte{
+						tls.PointFormatUncompressed,
+					}},
+
+					&tls.SessionTicketExtension{},
+					&tls.ALPNExtension{AlpnProtocols: []string{"h2", "http/1.1"}},
+					&tls.StatusRequestExtension{},
+					&tls.DelegatedCredentialsExtension{
+						AlgorithmsSignature: []tls.SignatureScheme{
+							tls.ECDSAWithP256AndSHA256,
+							tls.ECDSAWithP384AndSHA384,
+							tls.ECDSAWithP521AndSHA512,
+							tls.ECDSAWithSHA1,
+						},
+					},
+					&tls.KeyShareExtension{[]tls.KeyShare{
+						{Group: tls.X25519},
+						{Group: tls.CurveP256},
+					}},
+					&tls.SupportedVersionsExtension{[]uint16{
+						tls.VersionTLS13,
+						tls.VersionTLS12,
+					}},
+					&tls.SignatureAlgorithmsExtension{SupportedSignatureAlgorithms: []tls.SignatureScheme{
+						tls.ECDSAWithP256AndSHA256,
+						tls.ECDSAWithP384AndSHA384,
+						tls.ECDSAWithP521AndSHA512,
+						tls.PSSWithSHA256,
+						tls.PSSWithSHA384,
+						tls.PSSWithSHA512,
+						tls.PKCS1WithSHA256,
+						tls.PKCS1WithSHA384,
+						tls.PKCS1WithSHA512,
+						tls.ECDSAWithSHA1,
+						tls.PKCS1WithSHA1,
+					}},
+					&tls.PSKKeyExchangeModesExtension{[]uint8{
+						tls.PskModeDHE,
+					}},
+					&tls.FakeRecordSizeLimitExtension{0x4001},
+					&tls.UtlsPaddingExtension{GetPaddingLen: tls.BoringPaddingStyle},
+				}}, nil
+		},
+	},
+	settings: map[http2.SettingID]uint32{
+		http2.SettingHeaderTableSize:   65536,
+		http2.SettingInitialWindowSize: 131072,
+		http2.SettingMaxFrameSize:      16384,
+	},
+	settingsOrder: []http2.SettingID{
+		http2.SettingHeaderTableSize,
+		http2.SettingInitialWindowSize,
+		http2.SettingMaxFrameSize,
+	},
+	pseudoHeaderOrder: []string{
+		":method",
+		":path",
+		":authority",
+		":scheme",
+	},
+	connectionFlow: 12517377,
+	headerPriority: &http2.PriorityParam{
+		StreamDep: 13,
+		Exclusive: false,
+		Weight:    41,
+	},
+	priorities: []http2.Priority{
+		{StreamID: 3, PriorityParam: http2.PriorityParam{
+			StreamDep: 0,
+			Exclusive: false,
+			Weight:    200,
+		}},
+		{StreamID: 5, PriorityParam: http2.PriorityParam{
+			StreamDep: 0,
+			Exclusive: false,
+			Weight:    100,
+		}},
+		{StreamID: 7, PriorityParam: http2.PriorityParam{
+			StreamDep: 0,
+			Exclusive: false,
+			Weight:    0,
+		}},
+		{StreamID: 9, PriorityParam: http2.PriorityParam{
+			StreamDep: 7,
+			Exclusive: false,
+			Weight:    0,
+		}},
+		{StreamID: 11, PriorityParam: http2.PriorityParam{
+			StreamDep: 3,
+			Exclusive: false,
+			Weight:    0,
+		}},
+		{StreamID: 13, PriorityParam: http2.PriorityParam{
+			StreamDep: 0,
+			Exclusive: false,
+			Weight:    240,
+		}},
+	},
 }
 
 var Firefox_110 = ClientProfile{
