+++ tests/random_extension_order_test.go
@@ -3,6 +3,7 @@ package tests
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/bogdanfinn/tls-client/profiles"
 	"io"
 	"strings"
 	"testing"
@@ -15,7 +16,7 @@ import (
 
 func TestClient_RandomExtensionOrderChrome(t *testing.T) {
 	options := []tls_client.HttpClientOption{
-		tls_client.WithClientProfile(tls_client.Chrome_107),
+		tls_client.WithClientProfile(profiles.Chrome_107),
 		tls_client.WithRandomTLSExtensionOrder(),
 	}
 
@@ -65,7 +66,7 @@ func TestClient_RandomExtensionOrderChrome(t *testing.T) {
 
 func TestClient_RandomExtensionOrderCustom(t *testing.T) {
 	options := []tls_client.HttpClientOption{
-		tls_client.WithClientProfile(tls_client.CloudflareCustom),
+		tls_client.WithClientProfile(profiles.CloudflareCustom),
 		tls_client.WithRandomTLSExtensionOrder(),
 	}
