+++ tests/redirect_and_timeout_test.go
@@ -2,6 +2,7 @@ package tests
 
 import (
 	"fmt"
+	"github.com/bogdanfinn/tls-client/profiles"
 	"testing"
 	"time"
 
@@ -17,7 +18,7 @@ func TestClient_RedirectNoFollowWithSwitch(t *testing.T) {
 	defer testServer.Close()
 
 	options := []tls_client.HttpClientOption{
-		tls_client.WithClientProfile(tls_client.Chrome_105),
+		tls_client.WithClientProfile(profiles.Chrome_105),
 		tls_client.WithNotFollowRedirects(),
 	}
 
@@ -56,7 +57,7 @@ func TestClient_RedirectFollowWithSwitch(t *testing.T) {
 	defer testServer.Close()
 
 	options := []tls_client.HttpClientOption{
-		tls_client.WithClientProfile(tls_client.Chrome_105),
+		tls_client.WithClientProfile(profiles.Chrome_105),
 	}
 
 	client, err := tls_client.NewHttpClient(tls_client.NewNoopLogger(), options...)
@@ -94,7 +95,7 @@ func TestClient_TestFailWithTimeout(t *testing.T) {
 	defer testServer.Close()
 
 	options := []tls_client.HttpClientOption{
-		tls_client.WithClientProfile(tls_client.Chrome_105),
+		tls_client.WithClientProfile(profiles.Chrome_105),
 		tls_client.WithTimeoutSeconds(3),
 	}
